refix dataframe export
