+++ r/sedonadb/src/rust/src/dataframe.rs
@@ -28,7 +28,7 @@ use datafusion_expr::JoinType;
 use datafusion_expr::{select_expr::SelectExpr, Expr, SortExpr};
 use datafusion_ffi::table_provider::FFI_TableProvider;
 use savvy::{savvy, savvy_err, sexp, IntoExtPtrSexp, Result};
-use sedona::dataframe::{SedonaDataFrame, SedonaWriteOptions};
+use sedona::context::{SedonaDataFrame, SedonaWriteOptions};
 use sedona::show::{DisplayMode, DisplayTableOptions};
 use sedona_extension::streaming::StreamingRecordBatchReader;
 use sedona_geoparquet::options::TableGeoParquetOptions;
